==> contributors/aliciapouw21/1_Post_LidarFusion.ipynb <==
## After GPR-Lidar Fusion

import datetime
import pandas as pd
import numpy as np
import geopandas as gpd
import matplotlib.pyplot as plt

## Identify the multiple GPR points within LiDAR cells and find median TWTs
 - be sure to update CSV name
 - remoce random IX and lidar SD column when proper data updated
 - update column names 
 
 
 - Column names:
     - ASO_ID
     - ASO_X
     - ASO_Y
     - ASO_2020_02_02_SD
     - ASO_2020_02_02_SWE
     - ASO_2020_02_13_SD
     - ASO_2020_02_13_SWE
     
     - twt
     - 

#import data change name based on csv file
dataFusion= pd.read_csv("GPR_AND_ASO_ALL_FIXED.csv")
dataFusion = gpd.GeoDataFrame(dataFusion)
dataFusion.geometry = gpd.points_from_xy(dataFusion["ASO_X"], dataFusion["ASO_Y"]) # set to GPR x,y right now
print(dataFusion.dtypes) # check that floats are floats
dataFusion.head(10000).plot() # should see spiral

dataFusion["date"] = pd.to_datetime(dataFusion["date"])
#dataFusion = geopandas.GeoDataFrame(dataFusion, geometry = geopandas.points_from_xy(dataFusion.ASO_X, dataFusion.ASO_Y))
# Make lidar X,Y the geom column 

# %whos - this will tell you what variables you have 

#filter by date to seperate 2 data sets
date1 = '2020-02-02'
date2 = '2020-02-04'
dataFusion['date'] = pd.to_datetime(dataFusion['date'])
GPRLiDAR1 = dataFusion[(dataFusion['date'] <= date1)] #less than and equal to
GPRLiDAR2 = dataFusion[(dataFusion['date'] >= date2)]  #greater than and equal to this date 
print(len(GPRLiDAR1), ',', len(GPRLiDAR2))



GPRLiDAR2.head()
#len(GPRLiDAR)

dataFusion.head()

Data specific for date analysis only

##need to look at !!!

##Filtering by date for analysis on dates -- 
GPRLiDARdate1 = GPRLiDAR1.sort_values(by ='ASO_IDX', ascending = True)
GPRLiDARdate1 = GPRLiDAR1.groupby(['ASO_IDX', 'date']).agg({'twt': 'median', 'ASO_X': 'first', 'ASO_Y': 'first', 'ASO_2020_02_02_SD': 'first', 'ASO_2020_02_02_SWE': 'median'})
GPRLiDARdate1 = GPRLiDARdate1.sort_values(by ='ASO_IDX', ascending = True)

length = len(GPRLiDARdate1)
print(length)
GPRLiDARdate1.head()


GPRLiDARdate2 = GPRLiDAR2.sort_values(by ='ASO_IDX', ascending = True)
GPRLiDARdate2 = GPRLiDARdate2.groupby(['ASO_IDX', 'date']).agg({'twt': 'median', 'ASO_X': 'first', 'ASO_Y': 'first', 'ASO_2020_02_13_SD': 'first', 'ASO_2020_02_13_SWE': 'median'})
length = len(GPRLiDARdate2)
print(length)

GPRLiDARdate2.head()

Data we have grouped fully based on lidar cell!

##Filtering ignoring different dates. 
GPRLiDAR1 = GPRLiDAR1.sort_values(by ='ASO_IDX', ascending = True)
GPRLiDAR1 = GPRLiDAR1.groupby(['ASO_IDX']).agg({'twt': 'median', 'ASO_X': 'first', 'ASO_Y': 'first', 'ASO_2020_02_02_SD': 'first', 'ASO_2020_02_02_SWE': 'median', 'date': 'mean'})

GPRLiDAR2 = GPRLiDAR2.sort_values(by ='ASO_IDX', ascending = True)
GPRLiDAR2 = GPRLiDAR2.groupby(['ASO_IDX']).agg({'twt': 'median', 'ASO_X': 'first', 'ASO_Y': 'first', 'ASO_2020_02_13_SD': 'first', 'ASO_2020_02_13_SWE': 'median', 'date': 'mean'})

length = len(GPRLiDAR2)

print([length, len(GPRLiDAR1)])

GPRLiDAR1.head()

## Calculate snow density 
- eq 1. hs = TWT(v)/2  hs = Lidar snow depth, TWT = gpr two way travel time, V = wave speed
- eq 2. V = C/ sqrt(e),  V = wave speed, C = speed of light (0.3m/ns), e = dielectric permittivity
- eq 3. e = (1+0.845d)^2  Kovacs et al. 1995 d = density (kg/m3 /1000)

re-arrange eq 1.
- V = hs(2)/TWT

re-arrange eq 2.
- e = C^2/V^2

re-arrange eq 3.
- d = (sqrt(e)-1)/0.845

- update lidarSD column name 

#First calculation 1
#V = wave speed
GPRLiDAR1['v']= (GPRLiDAR1["ASO_2020_02_02_SD"]*2)/GPRLiDAR1["twt"]
GPRLiDAR2['v']= (GPRLiDAR2["ASO_2020_02_13_SD"]*2)/GPRLiDAR2["twt"]

## date specific data set
GPRLiDARdate1['v']= (GPRLiDARdate1["ASO_2020_02_02_SD"]*2)/GPRLiDARdate1["twt"]
GPRLiDARdate2['v']= (GPRLiDARdate2["ASO_2020_02_13_SD"]*2)/GPRLiDARdate2["twt"]

#second Calculation 
#e = dielectric permitivity
GPRLiDAR1['e'] = 0.3**2/(GPRLiDAR1['v']**2)
GPRLiDAR2['e'] = 0.3**2/(GPRLiDAR2['v']**2)

## date specific data set
GPRLiDARdate2['e'] = 0.3**2/(GPRLiDARdate2['v']**2)
GPRLiDARdate1['e'] = 0.3**2/(GPRLiDARdate1['v']**2)

#third calculation
# density
# d = (sqrt(e)-1)/0.845
GPRLiDAR1['dens'] = (((GPRLiDAR1['e'])**(1/2))-1)/0.845
#convert units to kg/m^3
GPRLiDAR1['dens'] = GPRLiDAR1['dens']*1000

GPRLiDAR2['dens'] = (((GPRLiDAR2['e'])**(1/2))-1)/0.845
#convert units to kg/m^3
GPRLiDAR2['dens'] = GPRLiDAR2['dens']*1000

##date specific dataset
GPRLiDARdate1['dens'] = (((GPRLiDARdate1['e'])**(1/2))-1)/0.845
#convert units to kg/m^3
GPRLiDARdate1['dens'] = GPRLiDARdate1['dens']*1000

GPRLiDARdate2['dens'] = (((GPRLiDARdate2['e'])**(1/2))-1)/0.845
#convert units to kg/m^3
GPRLiDARdate2['dens'] = GPRLiDARdate2['dens']*1000

GPRLiDAR1.head()

## Simple Median Filter to remove 
- running median filter using a buffer


import scipy.spatial as spatial

points = np.array((GPRLiDAR1["ASO_X"],GPRLiDAR1["ASO_Y"]))
points = points.transpose()
print(points)
point_tree = spatial.cKDTree(points)
buffer = 25 ## 25 meters
out = point_tree.query_ball_point(points, 25)

denseout = []
for x in out:
    denseout.append(GPRLiDAR1.iloc[x]['dens'].median())
len(denseout)

GPRLiDAR1['dens'] = denseout

points = np.array((GPRLiDAR2["ASO_X"],GPRLiDAR2["ASO_Y"]))
points = points.transpose()
print(points)
point_tree = spatial.cKDTree(points)
buffer = 25 ## 25 meters
out = point_tree.query_ball_point(points, 25)

denseout = []
for x in out:
    denseout.append(GPRLiDAR2.iloc[x]['dens'].median())
len(denseout)

GPRLiDAR2['dens'] = denseout


## Calculate SWE using 
- GPR density (kg/m3)
- LiDAR Snow Depth (m)

- SWE (m) = LiDAR SD (m) * GPR Density (kg/m3)

## all dates of same location between two time frames data 
GPRLiDAR1['SWE'] = (GPRLiDAR1['ASO_2020_02_02_SD']) * (GPRLiDAR1['dens']/1000)
GPRLiDAR2['SWE'] = GPRLiDAR2['ASO_2020_02_13_SD'] * (GPRLiDAR2['dens']/1000)

##dates only filtered by
GPRLiDARdate1['SWE'] = GPRLiDARdate1['ASO_2020_02_02_SD'] * (GPRLiDARdate1['dens']/1000)
GPRLiDARdate2['SWE'] = GPRLiDARdate2['ASO_2020_02_13_SD'] * (GPRLiDARdate2['dens']/1000)

GPRLiDAR1.head()

## ASO calculated density from Lidar-GPR fusion and ASO SWE

$\rho = \frac{SWE}{SD}$


GPRLiDAR1['ASO_2020_02_02_dens'] = (GPRLiDAR1.ASO_2020_02_02_SWE / GPRLiDAR1.ASO_2020_02_02_SD)*1000
GPRLiDARdate1['ASO_2020_02_02_dens'] = (GPRLiDARdate1.ASO_2020_02_02_SWE / GPRLiDARdate1.ASO_2020_02_02_SD)*1000

GPRLiDARdate2['ASO_2020_02_13_dens'] = (GPRLiDARdate2.ASO_2020_02_13_SWE / GPRLiDARdate2.ASO_2020_02_13_SD)*1000
GPRLiDAR2['ASO_2020_02_13_dens'] = (GPRLiDAR2.ASO_2020_02_13_SWE / GPRLiDAR2.ASO_2020_02_13_SD)*1000

GPRLiDAR2.isna().sum()

## Export CSV with new values
Table should include:
- ASO X 
- ASO Y
- ASO snow depth
- ASO SWE
- GPR TWT
- GPR Density
- GPR/LiDAR SWE



GPRLiDAR1

## GPRLiDAR1 = January 28th - February 2nd ( Flight February 2nd 2020)
GPRLiDAR1.to_csv('GPRLiDAR1.csv')
## GPRLiDAR2 = Febraury 4 - February 9 (Flight February 13th 2020
GPRLiDAR2.to_csv('GPRLiDAR2.csv')

##filtered by dates and lidar cell location
GPRLiDARdate1.to_csv('GPRLiDARdate1.csv')
GPRLiDARdate2.to_csv('GPRLiDARdate2.csv')
