==> contributors/aliciapouw21/3_SnoTel_forGPR.ipynb <==
#Snotel data

#This is the latest CUAHSI API endpoint
wsdlurl = 'https://hydroportal.cuahsi.org/Snotel/cuahsi_1_1.asmx?WSDL'

import os
from datetime import datetime
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd
import geopandas as gpd
from shapely.geometry import Point
import contextily as ctx
import ulmo

sites = ulmo.cuahsi.wof.get_sites(wsdlurl)
